updated last of private capabilities
