+++ change/@microsoft-teams-js-c23ceb97-a527-4ef9-8714-f4beff7446ec.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Made the `copilot`, `hostEntity`, `messageChannels`, and `teams` files treeshakable",
+  "packageName": "@microsoft/teams-js",
+  "email": "109628470+noahdarveau-MSFT@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
