+++ packages/teams-js/src/private/copilot/copilot.ts
@@ -1,0 +1,3 @@
+import * as eligibility from './eligibility';
+
+export { eligibility };
