+++ packages/teams-js/src/private/index.ts
@@ -22,7 +22,7 @@ export * as conversations from './conversations';
 //It is necessary to export OpenConversationRequest from conversations.ts individually as well
 //to keep the named exports so as to not break the existing consumers directly referencing the named exports.
 export { OpenConversationRequest } from './conversations';
-export { copilot } from './copilot';
+export * as copilot from './copilot/copilot';
 export * as externalAppAuthentication from './externalAppAuthentication';
 export * as externalAppAuthenticationForCEA from './externalAppAuthenticationForCEA';
 export * as externalAppCardActions from './externalAppCardActions';
@@ -30,11 +30,11 @@ export * as externalAppCardActionsForCEA from './externalAppCardActionsForCEA';
 export * as externalAppCommands from './externalAppCommands';
 export * as files from './files';
 export * as meetingRoom from './meetingRoom';
-export { messageChannels } from './messageChannels';
+export * as messageChannels from './messageChannels/messageChannels';
 export * as notifications from './notifications';
 export * as otherAppStateChange from './otherAppStateChange';
 export * as remoteCamera from './remoteCamera';
 export * as appEntity from './appEntity';
-export { teams } from './teams';
+export * as teams from './teams/teams';
 export * as videoEffectsEx from './videoEffectsEx';
-export { hostEntity } from './hostEntity';
+export * as hostEntity from './hostEntity/hostEntity';
