+++ packages/teams-js/test/private/hostEntity.spec.ts
@@ -1,6 +1,6 @@
 import { errorLibraryNotInitialized } from '../../src/internal/constants';
 import { GlobalVars } from '../../src/internal/globalVars';
-import { hostEntity } from '../../src/private/hostEntity';
+import * as hostEntity from '../../src/private/hostEntity/hostEntity';
 import { ErrorCode, FrameContexts } from '../../src/public';
 import * as app from '../../src/public/app/app';
 import { _minRuntimeConfigToUninitialize } from '../../src/public/runtime';
