+++ packages/teams-js/test/private/messageChannels.spec.ts
@@ -1,6 +1,6 @@
 import * as communication from '../../src/internal/communication';
 import { errorLibraryNotInitialized } from '../../src/internal/constants';
-import { messageChannels } from '../../src/private/messageChannels';
+import * as messageChannels from '../../src/private/messageChannels/messageChannels';
 import { FrameContexts } from '../../src/public';
 import * as app from '../../src/public/app/app';
 import { errorNotSupportedOnPlatform } from '../../src/public/constants';
